Crate mktape

mktape stems from the need to create a PDP-10 tap file to load source of another project of mine, onto an emulated PDP-10. As part of the research into making this work I was pointed to this paper and, in particular, the perl script referenced in Appendix A. The perl script creates a pre-named tap file (v7.tap) from files f0..f6. More over, the input files have pre-defined block sizes for each of them.

This project generalises the script so that the user can:

specify the output file (required)

specify the input file(s), minimally one file

specify the input file’s block-sizes (defaults to 1024 to align with the tapewrite default from the tapeutils repository).

Modules

mktape

result